backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
alembic
aiogram
sqlmodel
beautifulsoup4
lxml
